Move frustum intersect function
